requirements.txt
# ----------------------------------------------------------------
# PIP Dependencies, install using the following:
# $ pip install -r requirements.txt
# ----------------------------------------------------------------

sqlalchemy
mysqlclient
flask
flask_cors
gunicorn
python-dotenv
requests
flask-swagger-ui
flasgger
alembic
parameterized
httplib2
celery
redis
jinja2
bcrypt
pyjwt